requirements.txt
fastapi
httpx
pydantic
pytest
python-dotenv
python_jose
Requests
SQLAlchemy
uvicorn
asyncmy
cryptography